requirements.txt
python-dotenv
requests
fastapi
uvicorn
bs4